include/sycl/algorithm/ 	- sort.hpp 		replace int by size_t in lines 87, 88 	- exclusive_scan.hpp 		replace int by size_t in lines 114, 115 	- inner_product.hpp 		remove lines 128, 129 : unused variables (localid, globalid) 	- tranform_reduce.hpp 		remove line 66 : unused type 		remove lines 85, 86 : unused variables (localid, globalid) 	- inclusive_scan.hpp 		replace int by size_t in lines 92, 93 	- count_if.hpp 		remove line 66 : unused type 		remove lines 86, 87 : unused variables (localid, globalid) 	- for_each.hpp 		remove line 53 : unused type

+++ include/sycl/algorithm/sort.hpp
@@ -84,8 +84,8 @@ class sort_kernel_sequential {
 
   // Simple sequential sort
   void operator()() {
-    for (int i = 0; i < vS_; i++) {
-      for (int j = 1; j < vS_; j++) {
+    for (size_t i = 0; i < vS_; i++) {
+      for (size_t j = 1; j < vS_; j++) {
         if (a_[j - 1] > a_[j]) {
           sort_swap<T>(a_[j - 1], a_[j]);
         }
